limit notification update frequency to once per second

+++ app/src/main/java/me/hexian000/masstransfer/io/ProgressReporter.java
@@ -1,5 +1,5 @@
 package me.hexian000.masstransfer.io;
 
 public interface ProgressReporter {
-	void report(String text, int now, int max);
+	void report(String text, long now, long max);
 }
